refactor: replace Chromsizes with GenomicScale
