+++ hg/__init__.py
@@ -3,14 +3,15 @@ try:
 except ImportError:
     __version__ = "unknown"
 
-from higlass_schema import *
+# must import before hg since we extend higlass_schema behavior
+from higlass_schema import * # isort: skip
 
 import hg.tilesets
-from hg.api import *  # overrides classes with same name from higlass_schema
-from hg.chromsizes import Chromsizes
+from hg.api import *
 from hg.fuse import fuse
 from hg.server import server
 from hg.tilesets import remote
+from hg.utils import GenomicScale
 
 bigwig = server.register(hg.tilesets.bigwig)
 multivec = server.register(hg.tilesets.multivec)
